phc2sys: Add multi-domain synchronization.
Add support for synchronization between clocks controlled by different
ptp4l instances in the automatic mode (-a option) to allow switching to
a backup source on failures.

The -z and -n options can be used multiple times to specify sockets and
domain numbers of different ptp4l instances (up to 16 instances).

Clocks from one ptp4l instance are considered a domain. If the domain
doesn't have a source clock (i.e. a port in the slave state), the clocks
in the domain can be synchronized to a source clock from a different
domain. If multiple domains have a source clock, the first one in the
order of specified sockets is selected. The selection can be later
extended to compare some specific clock attributes.

The real-time clock is a separate domain, which is created only with the
-r option and can be a source for other domains if the -r option is
specified twice.

In the non-automatic modes only one domain is created for all clocks.

Signed-off-by: Miroslav Lichvar <mlichvar@redhat.com>