phc2sys: Improve logging with single domain.
When phc2sys in the automatic mode has only one non-CLOCK_REALTIME
domain, avoid logging the potentially confusing "selecting ... as
out-of-domain source clock" message. The source clock in the domain is
already logged. Also, log a message when CLOCK_REALTIME is selected for
synchronization to match the behavior of phc2sys before it supported
multiple domains.

Signed-off-by: Miroslav Lichvar <mlichvar@redhat.com>