Avoid conflicting port IDs over PMC UDS
The UDS interface is currently associated with a clock ID of zeros
and a 16-bit port number which is the process id. However, the
process id on Linux can easily by bigger than 16 bits (supposed
to be limited to 22 bits). This means that several linuxptp processes
(pmc and phc2sys for instance) can collide and use the same port id.

Hence, use the lower 2 bytes of the process id for the port number,
and the top 2 bytes as bytes 7 and 8 of the Clock ID.

Reviewed-by: Jacob Keller <jacob.e.keller@intel.com>
Signed-off-by: Eyal Itkin <eitkin@nvidia.com>