Make allowedLostResponses configurable
This change defines 'allowedLostResponses' as a per-port parameter
with a default value of 3 (per IEEE 802.1AS-2011, clause 11.5.3).
(Note that this also matches the value of the previously #define'd
ALLOWED_LOST_RESPONSES).

It is now configurable to comply with IEEE 802.1AS-2020, clause 11.5.3
which specifies a default value of 9 (and in the range of 1-255).

Signed-off-by: Kishen Maloor <kishen.maloor@intel.com>
Signed-off-by: Chwee-Lin Choong <chwee.lin.choong@intel.com>