phc2sys: Stop synchronization when ptp4l stops responding.
phc2sys in the automatic mode waits for ptp4l on start, but doesn't care
if it stopped providing state notifications and responding to UTC offset
requests, e.g. if it stopped, crashed, freezed, or its socket was
removed by a misconfigured ptp4l instance.

To better handle broken communication with ptp4l and make it evident,
if the pmc subscription runs out, log an error message and stop
synchronization in the domain until the pmc agent is subscribed again.

Signed-off-by: Miroslav Lichvar <mlichvar@redhat.com>