Fix invalid rogue peer delay request error
When linuxptp goes from LISTENING to UNCALIBRATED, it will flush the the peer delay.
This means that linuxptp forgets it sent a peer delay request.
However, if there is an outstanding peer delay request at this point in time,
the subsequent response will be considered rogue, and linuxptp will go into a faulty state.
Which causes it to do nothing for a while.

To fix this, we keep track of wether a flush_peer_delay has occurred.
And if we receive a response after one, we don't consider it rogue but just ignore it.

Signed-off-by: Joris Guisson <joris.guisson@luminex.be>