interface: Add an optional remote address for use by the UDS transport.
Because of the implementation of the UDS module, it is not possible
for a process to act as both a PTP management client (PMC) and server.
Going forward, one ptp4l instance will need to subscribe to another
using the PMC methods.  Pave the way by allowing the interface to
include an optional remote UDS address.

Signed-off-by: Richard Cochran <richardcochran@gmail.com>