ts2phc: Don't switch system clock to nanosecond mode.
ts2phc is not synchronizing the system clock and should not switch the
clock to the nanosecond mode with adjtimex(modes=ADJ_NANO) or make any
other modifications to it. The process that is controlling the clock
(e.g. an NTP client) might not be using the nanosecond mode.

There are two instances of the adjtimex() call in the code. One is used
only to read the clock and can be replaced with faster clock_gettime().
The other instance is also reading the TAI offset. Instead of switching
to the nanosecond mode, change the timestamp conversion to handle both
microsecond and nanosecond modes according to the current clock status.

Reviewed-by: Jacob Keller <jacob.e.keller@intel.com>
Signed-off-by: Miroslav Lichvar <mlichvar@redhat.com>