uds: Warn if replacing existing socket.
When running multiple ptp4l instances, they need to be configured with
different UDS addresses in order for pmc/phc2sys to be able to
communicate with all instances. If this is not the case, the socket is
silently replaced in the order of started ptp4l instances and they
cannot be all reached over UDS.

Log a warning message if the socket is removed to make it more obvious
that something is wrong.

Signed-off-by: Miroslav Lichvar <mlichvar@redhat.com>