Clean up compiler warning in port.c.
v2 - reformat commit message

Gcc (and presumably other compilers) emits a warning about the
declaration of 'dscmp' shadowing a global declaration in bmc.h.
This warning is completely benign but makes it more difficult to
recognize warnings that might be consequential.

This patch renames the variable 'dscmp' to 'clk_dscmp'.

Signed-off-by: Michael Galassi <michael@galassi.us>