Clean up compiler warning in ts2phc.c.
v2 - reformat commit message

Gcc (and presumably other compilers) emits a warning about the
declaration of 'c' shadowing a previous local in ts2phc.c.  This
warning is completely benign but makes it more difficult to recognize
warnings that might be consequential.

This patch renames the variable 'c' to 'clk'.

Signed-off-by: Michael Galassi <michael@galassi.us>