Clean up compiler warning in ts2phc_pps_sink.c.
v2 - reformat commit message

Gcc (and presumably other compilers) emits a warning about the
declaration of 'sink' shadowing a previous local in ts2phc_pps_sink.c.
This warning is completely benign but makes it more difficult to
recognize warnings that might be consequential.

This patch reuses the previous declaration of sink which is also a
pointer to a struct ts2phc_pps_sink.

Signed-off-by: Michael Galassi <michael@galassi.us>