ts2phc: Add holdover support.
If the external PPS signal is generated by a clock with better long-term
stability than the clock synchronizing to it, it can work as a good time
source for some period of time after losing its own time source (e.g.
GPS receiver losing its signal).

Add an option to specify a holdover interval where ts2phc can continue
synchronizing the clock without any ToD information. Allow that only in
the LOCKED_STABLE servo state, which needs to be enabled by the
servo_num_offset_values option.

This is supported only in the non-automatic mode and when the pulse
polarity detection is disabled.

Signed-off-by: Miroslav Lichvar <mlichvar@redhat.com>