sad: introduce sad_append_auth_tlv() and sad_generate_icv()
sad_append_auth_tlv() will append an authentication tlv as specified by
the security association corresponding to the passed spp. After
attached, msg_pre_send() is called to put message in network byte order
and sad_generate_icv() is called to populate the icv field in the tlv.

the allow_mutable flag will set the correction field to 0 before
generating the icv to allow for configurations where on path auth tlv
support is not present.

Signed-off-by: Clay Kaiser <Clay.Kaiser@ibm.com>
Reviewed-by: Erez Geva <ErezGeva2@gmail.com>
Reviewed-by: Miroslav Lichvar <mlichvar@redhat.com>