phc2sys: Keep clocks in command-line or ptp4l order.
When adding a new clock to the domain, add it to the end of the list to
keep them in the same order as specified on the command line or the port
order of ptp4l.

This will be needed by new code in the domain reconfiguration expecting
the order of reinitialized clocks to be the same as in which they were
added to the domain.

Signed-off-by: Miroslav Lichvar <mlichvar@redhat.com>
Reviewed-by: Jacob Keller <jacob.e.keller@intel.com>