phc2sys: Allow static sink clocks in automatic mode.
Allow the -c option to be used together with the -a option. Add the
specified clocks to the first domain, where they will stay in the master
state and be synchronized when a source is available in the same domain
or other domains.

Mark the clocks added by -c as static and skip them in the domain
reconfiguration if they duplicate a clock following a ptp4l port.

A use case is synchronization of clocks of backup interfaces in an
active-backup bond to minimize the observed offset when the active
interface is switched. ptp4l tracks the active interface, which is
followed by the automatic mode of phc2sys. All interfaces included in
the bond are specified by the -c option to keep them all synchronized to
the same source. The interface which duplicates the current active
interface provided by ptp4l is skipped.

Signed-off-by: Miroslav Lichvar <mlichvar@redhat.com>