port: add port configuration option for profile identifier
This commit adds configuration of 6-octet profile identifiers for use
in the clock description. The option helps identify profiles that may
not always have configurations unique from the default or other profiles.

When using the default 00:00:00:00:00:00 or when not defined, the clock
description falls back to the existing profile checks.

Signed-off-by: William Comly <william.comly@nav-timing.safrangroup.com>