Fix man page installation for out of tree builds.
The makefile supports out of tree builds by using the GNU make VPATH
facility.  However, the `make install` target fails to install the man
pages because, unlike the generated binaries, the man page files exist
only in the source directory and not in the build directory.  Fix the
issue by adding the source path prefix to the `install` command line.

Reported-by: Martin Pecka <peckama2@fel.cvut.cz>
Signed-off-by: Richard Cochran <richardcochran@gmail.com>