phc2sys: Don't require -O option without -a and -w.
Allow phc2sys to be started in the static mode without the -O option,
meaning it should fetch the UTC offset from ptp4l, even if the -w option
is not included on the command line to wait for ptp4l to have a port in
the master or slave state.

This is useful when the PHC is synchronized by ts2phc or other
application and the system clock needs to be synchronized to the PHC
even when ptp4l has all ports in the faulty state (e.g. all network
links are down).

It's the responsibility of the user to start phc2sys when ptp4l is
already running. phc2sys will exit with an error message if it cannot
get a response with the 1-second timeout.

Signed-off-by: Miroslav Lichvar <mlichvar@redhat.com>