clock: Add variables for external grandmaster.
When ptp4l is configured to run as the server part of a bridge between
different domains, or a boundary clock split into multiple instances
(using one or two clocks), it announces its own clock identity as
grandmasterIdentity and stepsRemoved of 0. The clients of the server
don't see the true identity and distance, which can impact the result of
BMCA if there are multiple bridges/paths to the grandmaster. In the
worst case it could cause synchronization loops.

Add new variables to the clock that will override the grandmaster
identity and stepsRemoved values included in announce messages to make
it possible to pass them from the source domain as if the bridge/BC was
a single PTP instance. The variables will be used only when in the
grandmaster state and when not zero.

Runtime configuration over the UDS port will follow.

Signed-off-by: Miroslav Lichvar <mlichvar@redhat.com>