Add management message for timestamp and delay corrections.
Add a management message to get and set the port egressLatency,
ingressLatency and delayAsymmetry variables. This enables an external
process to compensate for changes in the network and timestamping
asymmetries while ptp4l is running.

Exchange the values as fractional nanoseconds, but parse and print them
in pmc as whole nanoseconds.

Signed-off-by: Miroslav Lichvar <mlichvar@redhat.com>