ts2phc: Add support for higher pulse rates.
1 pulse per second might be insufficient with some specific requirements
for clocks synchronized by ts2phc. Analogously to the phc2sys -R option,
add ts2phc.pulserate option to enable synchronization with higher pulse
rates. The specified rate modifies the calculation of the source
timestamp needed for the sink offset and also the PEROUT_REQUEST for the
PHC PPS source if enabled.

To make the configuration easier, change the default pulsewidth to be
50% of the pulse period instead of fixed 0.5 seconds. Exit with an error
message if the specified width is too large for the given rate.

Signed-off-by: Miroslav Lichvar <mlichvar@redhat.com>
Reviewed-by: Jacob Keller <jacob.e.keller@intel.com>