ts2phc: Add option for external reference in automatic mode.
ts2phc running in the automatic mode assumes that the PPS source is one
of the PHCs used by ptp4l. That doesn't always have to be the case. Add
"external_pps" option to use the non-automatic-mode reading of the
source timestamp and don't mark sink clocks as targets if they are
synchronized by ptp4l.

The use case is holdover with an externally controlled stabilized clock.
The clock is synchronized by an external process to the PHC when it's
synchronized by ptp4l and the PHC is synchronized to the external clock
by ts2phc when not synchronized by ptp4l. Multiple PHCs can be connected
to the external clock to make a JBOD boundary clock.

Don't require the PHC that is synchronized by ptp4l to be receiving the
PPS signal (providing timestamps) to make it possible to switch the PPS
direction to synchronize the external clock if there is only one
usable channel.

Signed-off-by: Miroslav Lichvar <mlichvar@redhat.com>
Reviewed-by: Jacob Keller <jacob.e.keller@intel.com>