Add ENABLE_PORT and DISABLE_PORT commands.
Add support for the ENABLE_PORT and DISABLE_PORT management
commands defined in 1588 to allow enabling/disabling individual ports
without having to restart ptp4l.

The code already supports the DESIGNATED_ENABLED and DESIGNATED_DISABLED
events that are specified for the transition from/to the DISABLED port
state.

Add the command action to pmc and trigger the events in the port when
the commands are received. The commands are accepted only on the UDS-RW
port.

Signed-off-by: Miroslav Lichvar <mlichvar@redhat.com>