util: Add functions for dropping root privileges.
Add a function to switch the process UID/GID to a specified user in
order to drop the root privileges. Keep only the capabilities needed to
adjust the clock, enable HW timestamping, bind to privileged ports and
raw sockets, using the libcap library. Initialize the supplementary
groups of the user to make it easier to give it permissions to open PHC
devices by adding it to a system group.

Add a function to create a directory for a UDS address with a specified
owner, where the programs will be able to bind and unlink their sockets
without root privileges.

Signed-off-by: Miroslav Lichvar <mlichvar@redhat.com>
Reviewed-by: Jacob Keller <jacob.e.keller@intel.com>