pmc_agent: Keep pmc_ds_requested set across run_pmc() calls.
In the handling of timeout in run_pmc() the pmc_ds_requested variable is
always cleared to force the next call of the function to send a new
request if a response is not received in that time. This prevents
detection of a missed response as nothing is tracking if a request was
sent in the previous call. The variable is also set in run_pmc_wait_sync()
to avoid sending a new request when waiting multiple port-specific
responses.

Modify the function to accept a new parameter enabling the receive-only
mode and keep the pmc_ds_requested variable set on the timeout to make
it clear when a response is missing, which will be needed in the next
commit to reopen the transport.

Signed-off-by: Miroslav Lichvar <mlichvar@redhat.com>
Reviewed-by: Jacob Keller <jacob.e.keller@intel.com>