pmc_agent: Reopen transport after missed response.
If the pmc agent (used in phc2sys and ts2phc) was started as root before
ptp4l configured to not run as root could bind its socket, the client
socket owner was not changed to follow the server socket, which caused
the non-privileged server to not be able to send responses to the
client.

Add a new function to pmc_common to allow reopening the transport
without destroying the pmc instance, and call it in the agent when a
response from ptp4l is missed. This allows the client to correct the
socket owner if it was started too soon in order to receive responses
from the server. It also allows the client to recover if its socket is
inadvertently removed.

Signed-off-by: Miroslav Lichvar <mlichvar@redhat.com>
Reviewed-by: Jacob Keller <jacob.e.keller@intel.com>