raw: Prevent leakage of non-PTP packets during socket init
There were two problems with the socket configuration sequencing:

1) Calling socket(PF_PACKET, SOCK_RAW, htons(ETH_P_ALL)) causes all
ethernet frames from -all- interfaces to be queued to the socket,
until bind() is later called.

2) The BPF filter is installed -after- bind() is called, so ethernet
frames that should be rejected could be queued before the BPF filter
is installed.

This patch reorders the raw socket initialization so that all
configuration happens before bind() is called.

Signed-off-by: Cliff Spradlin <cspradlin@google.com>
Reviewed-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>